source/config/requirements/requirements-azure.txt
-r ${PWD}/config/requirements/base.txt

django-storages[azure]
